+++ src/constants/baseURL.js
@@ -1,3 +1,3 @@
-const BASE_URL = 'http://localhost:3001';
+const BASE_URL = 'http://localhost:3000';
 
 export default BASE_URL;
