+++ src/services/api.js
@@ -1,19 +1,5 @@
+import { request } from '../utils';
 import { BASE_URL } from '../constants';
-
-const request = async (url, option = {}) => {
-  try {
-    const res = await fetch(url, option);
-    const body = await res.json();
-
-    if (!res.ok) {
-      throw new Error(`http request Error : ${res.status}`);
-    }
-
-    return body;
-  } catch (error) {
-    throw new Error(`http request Error : ${error}`);
-  }
-};
 
 const API = {
   getItemList: async () => {
@@ -47,4 +33,4 @@ const API = {
   },
 };
 
-export default API;
+export default API
