+++ src/utils/index.js
@@ -1,1 +1,1 @@
-export { default as API } from './api';
+export { default as request } from './request';
